refactor(io): Make Drive URL detection reachable from _remote
Move the pure-stdlib `_is_gdrive_url` / `_GDRIVE_HOSTS` Drive-link detection
into `_remote.py` (alongside `_is_url`) and re-export them from `_gdrive.py`,
so detection is importable without pulling in the heavier resolver module and
is reachable as `_remote._is_gdrive_url` for the notebook-safety check. The
`open_url` Drive branch now uses the local detector and imports the
`_open_gdrive` resolver lazily only when a Drive URL is actually seen.

Add an `open_url`-level integration test that routes a `drive.google.com`
URL through the resolver against the loopback pytest-httpserver, covering the
Drive branch of `open_url` itself (no real Google Drive request).
